requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
cffi==1.14.5
channels==3.0.3
channels-redis==3.3.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Deprecated==1.2.13
Django==3.2.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kombu==5.1.0
msgpack==1.0.2
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.19
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.6.0
pyOpenSSL==20.0.1
pyparsing==3.0.9
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zope.interface==5.4.0
